### 2025-08-04 #8737 @editorialbot
[PRE REVIEW]: lammpsio: Transparent and reproducible handling of LAMMPS particle data in Python
**Submitting author:** @clpetix (<a href="http://orcid.org/0000-0002-0483-7495">C. Levi Petix</a>)
**Repository:** https://github.com/mphowardlab/lammpsio
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.8.0
