### 2025-08-26 #8737 @rkurchin
Thanks for your help here, Mehul! Since I believe this is your first time reviewing for JOSS, JFYI the review will officially start once I find at least one more reviewer, then there will be a transparent, checklist-driven process that happens over on the review issue. You can of course let me know if you have any questions along the way!
