### 2025-09-01 #8737 @rosecers
> 👋 [@rosecers](https://github.com/rosecers), [@aingaranking](https://github.com/aingaranking) and/or [@jpotoff](https://github.com/jpotoff), would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @rkurchin I would love to, but am a bit over-subscribed at the moment. I can review in ~1 month or so, but will not have time before then, unfortunately.
