### 2025-09-09 #8962 @editorialbot
[REVIEW]: lammpsio: Transparent and reproducible handling of LAMMPS particle data in Python
**Submitting author:** @clpetix (<a href="http://orcid.org/0000-0002-0483-7495">C. Levi Petix</a>)
**Repository:** https://github.com/mphowardlab/lammpsio
**Branch with paper.md** (empty if default branch): joss
**Version:** 0.9.0
