### 2025-09-12 mphowardlab/lammpsio#90 @Ladme
JOSS Review
Hello,

I'm reviewing your article and package for [JOSS](https://github.com/openjournals/joss-reviews/issues/8962).

Both the package and the article look good and pass the required JOSS checks. I particularly like the easy integration with other analysis packages like `freud`, the documentation and CI.

I have a few minor recommendations that I believe would further improve the `lammpsio` package and the documentation:

1) **Add type hints.**
The package is well-documented, but adding explicit type annotations would make it easier for users (and especially IDEs) to understand what types of inputs the functions expect and what they return.

2) **Link the documentation from the GitHub repository.**
The documentation is very nice, but currently not easy to find. I would suggest adding a link to the documentation either in the README or in the Website section of the repository details (or both).

3) **Make very small fixes in the tutorials.**
- In the *Analyzing a Lennard-Jones fluid* tutorial, block [3] includes `import numpy`, but this import is not used and is repeated in block [4]. I belive it can be safely removed from [3].
- The tutorials seem intended for use in a Jupyter Notebook. That’s fine, but it would help to make this explicit. When run as scripts, the examples do not work as shown -- e.g. `rdf.plot();` does not display a figure unless `matplotlib.pyplot` is also imported and the plot is explicitly saved.

Overall, this is a nice little package, and I’m sure it will be useful to its audience.
