### 2025-09-15 24548db
Add documentation to README.
