### 2025-09-15 a1732ff
Add type-hinting to Box.
