### 2025-09-15 2cacd56
Add type hints to DumpFile classes
