### 2025-09-15 f9edd20
Add type hints to Snapshot
