### 2025-09-15 41eb007
Make sure sequence of floats can also be ndarray.
