### 2025-09-18 3d5a5f8
Use Union() in place of | operator.
