### 2025-09-22 a7eaf01
Implement Ladme's suggestions from JOSS Review (#91)
* Add documentation to README.

* Make plotting compatible with plain Python.

* Add type-hinting to Box.

* Add type hints to data.py

* Add type hints to DumpFile classes

* Add type hints to Snapshot

* Make sure sequence of floats can also be ndarray.

* Add type hints to topology.py.

* Use Union() in place of | operator.

* Fix README typo.

* Remove typehints from documentation.
