+++ doc/source/conf.py
@@ -71,6 +71,7 @@ autosummary_generate = True
 autodoc_member_order = "bysource"
 
 autodoc_default_options = {"inherited-members": None, "special-members": False}
+autodoc_typehints = "none"
 
 # -- Options for intersphinx -------------------------------------------------
 
