### 2025-10-04 #8962 @clpetix
@rkurchin, thanks for staying on top of this! @Ladme replied on the pull request I linked [here](https://github.com/mphowardlab/lammpsio/pull/91#issuecomment-3318796125) and indicated that we had addressed all of his comments on [the Issue](https://github.com/mphowardlab/lammpsio/issues/90#issuecomment-3318798131) he opened in our repo. We just forgot to bring the conversation back to this thread.
