+++ setup.cfg
@@ -15,18 +15,18 @@ classifiers =
     License :: OSI Approved :: BSD License
     Topic :: Scientific/Engineering :: Chemistry
     Topic :: Scientific/Engineering :: Physics
-    Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
     Programming Language :: Python :: 3.11
     Programming Language :: Python :: 3.12
     Programming Language :: Python :: 3.13
+    Programming Language :: Python :: 3.14
 
 [options]
 packages = find:
 package_dir =
     = src
 include_package_data = True
-python_requires = >=3.9
+python_requires = >=3.10
 install_requires =
     numpy
     packaging
