### 2025-10-27 d66537a
Pre-commit and documentation fix
