+++ doc/source/tutorials.rst
@@ -6,6 +6,7 @@ Tutorials
 
     ./tutorials/dimer_lattice_tutorial
     ./tutorials/analysis_tutorial/analysis_tutorial
+    ./tutorials/numba_tutorial/numba_tutorial
 
 .. grid:: 2
 
@@ -22,3 +23,10 @@ Tutorials
 
         Read and analyze a LAMMPS dump file to study the Lennard-Jones fluid
         using the freud Python package.
+
+    .. grid-item-card:: Analyzing a 10-Bead Polymer with Numba
+        :link: ./tutorials/numba_tutorial/numba_tutorial
+        :link-type: doc
+
+        Read and analyze a LAMMPS dump file to study a 10-bead polymer
+        using the Numba library for accelerated computation.
