### 2025-11-07 #8962 @rkurchin
Note to myself that progress is ongoing in the linked issues (otherwise I keep inadvertently flagging this as "inactive" for 2+ weeks 😉).
