### 2025-11-07 5189e16
Remove initialization script and trajectory file
