### 2025-11-10 b52c1b0
Add tutorial for initializing an ethanol-water mixture to docs.
