### 2025-11-10 ffa9b09
Fix pre-commit and documentation errors
