### 2025-11-19 658b361
Add tutorial for initializing an ethanol-water.  (#97)
* Add tutorial for initializing an ethanol-water mixture to docs.

* Add jupyter notebook to create ethanol-water mixture

* Fix pre-commit and documentation errors

* Fix formatting of input file references in ethanol-water mixture tutorial

* Update tutorial to use molecule JSON.

* Read PACKMOL data in one file access.

* Fix typo in hydrogen type.

---------

Co-authored-by: mayukhkundu14 <79770572+mayukh33@users.noreply.github.com>
