+++ doc/source/tutorials.rst
@@ -6,6 +6,7 @@ Tutorials
 
     ./tutorials/dimer_lattice_tutorial
     ./tutorials/analysis_tutorial/analysis_tutorial
+    ./tutorials/ethanol-water-tutorial/ethanol_water_mixture
 
 .. grid:: 2
 
@@ -22,3 +23,9 @@ Tutorials
 
         Read and analyze a LAMMPS dump file to study the Lennard-Jones fluid
         using the freud Python package.
+
+    .. grid-item-card:: Initializing an Ethanol-Water Mixture
+        :link: ./tutorials/ethanol-water-tutorial/ethanol_water_mixture
+        :link-type: doc
+
+        Use Snapshot to create a LAMMPS data file for an ethanol-water mixture.
