### 2025-11-19 9460596
Added reverse mapping implementation to LabelMap with unit tests
