+++ tests/test_topology.py
@@ -204,16 +204,21 @@ def test_impropers_wrong_shape(snap_8):
 def test_LabelMap():
     # create a simple label map
     label = lammpsio.topology.LabelMap({1: "typeA", 2: "typeB"})
-
+    # check inverse mapping
+    assert label.inverse == {"typeA":1,"typeB":2}
     # check the types
     assert label.types == ("typeA", "typeB")
     # check the typeids
     assert label.typeid == (1, 2)
     # get
     assert label[2] == "typeB"
+    # get inverse
+    assert label.inverse["typeB"] == 2
     # set
     label[3] = "typeC"
     assert label[3] == "typeC"
+    assert label.inverse["typeC"] == 3
     # delete
     del label[3]
     assert 3 not in label
+    assert "typeC" not in label.inverse
