### 2025-11-19 a3a7c08
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
