+++ tests/test_topology.py
@@ -205,7 +205,7 @@ def test_LabelMap():
     # create a simple label map
     label = lammpsio.topology.LabelMap({1: "typeA", 2: "typeB"})
     # check inverse mapping
-    assert label.inverse == {"typeA":1,"typeB":2}
+    assert label.inverse == {"typeA": 1, "typeB": 2}
     # check the types
     assert label.types == ("typeA", "typeB")
     # check the typeids
