### 2025-11-19 ff686bf
Update LAMMPS input and tutorial files.
