+++ doc/source/tutorials/numba_tutorial/lammps_input.in
@@ -23,11 +23,7 @@ timestep 0.005
 fix 1 all nvt temp 1.0 1.0 1.0
 fix 2 all langevin 1.0 1.0 1.0 5
 
-# Output
-dump 1 all custom 1000 traj.lammpstrj id type x y z ix iy iz vx vy vz
-
 # Run simulation
-run 10000000
-undump 1
+run 10000
 dump 2 all custom 1000 traj.lammpstrj id type x y z ix iy iz vx vy vz
-run 100000000
+run 100000
