### 2025-11-19 7ae4407
Remove unnecessary line break in tutorials.rst
