### 2025-11-19 88a488b
Include multiprocessing example for reading snapshots in parallel
