+++ conftest.py
@@ -1,3 +1,5 @@
+import multiprocessing
+
 import numpy
 
 try:
@@ -30,6 +32,8 @@ def setup_sybil_tests(namespace):
     # Common imports.
     namespace["numpy"] = numpy
     namespace["lammpsio"] = lammpsio
+    namespace["multiprocessing"] = multiprocessing
+    namespace["Pool"] = multiprocessing.Pool
 
     if has_lammps:
         namespace["lammps"] = lammps
