+++ src/lammpsio/dump.py
@@ -125,7 +125,7 @@ class DumpFile:
 
         def process_snapshot(snapshot):
             return snapshot.N
-            
+
         if __name__ == '__main__':
             num_workers = max(4, multiprocessing.cpu_count())
             with multiprocessing.Pool(num_workers) as p:
