### 2025-11-19 f3ee0d5
Update src/lammpsio/topology.py
type hint usage

Co-authored-by: Michael Howard <mphoward@auburn.edu>
