### 2025-11-19 7dceb86
Apply suggestions from code review
Co-authored-by: Michael Howard <mphoward@auburn.edu>
