### 2025-11-19 5719b22
Final changes and credits addition
