### 2025-11-19 823d9bc
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
