### 2025-11-19 cac6195
Apply suggestion from @mphoward
