+++ doc/source/credits.rst
@@ -7,3 +7,7 @@ Credits
 * Philipp Leclercq
 * C\. Levi Petix
 * Dimitris Arquin
+* Michael P. Howard
+* Mayukh Kundu
+* Philipp Leclercq
+* C\. Levi Petix
