### 2025-11-19 e56f04c
Fix botched suggestion application
