+++ doc/source/credits.rst
@@ -2,10 +2,6 @@
 Credits
 =======
 
-* Michael P. Howard
-* Mayukh Kundu
-* Philipp Leclercq
-* C\. Levi Petix
 * Dimitris Arquin
 * Michael P. Howard
 * Mayukh Kundu
