### 2025-11-19 625d925
Added reverse mapping implementation to LabelMap  (#105)
* Added reverse mapping implementation to LabelMap with unit tests

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Update src/lammpsio/topology.py

type hint usage

Co-authored-by: Michael Howard <mphoward@auburn.edu>

* Apply suggestions from code review

Co-authored-by: Michael Howard <mphoward@auburn.edu>

* Added suggested documentation (became outdated)

* Final changes and credits addition

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see h
[…]
