+++ doc/source/credits.rst
@@ -2,6 +2,7 @@
 Credits
 =======
 
+* Dimitris Arquin
 * Michael P. Howard
 * Mayukh Kundu
 * Philipp Leclercq
