### 2025-11-19 bdce48e
Revise paper in response to reviewer feedback
