+++ paper/paper.md
@@ -29,7 +29,7 @@ lammpsio provides a Python interface for reading and writing particle data in LA
 
 # Statement of need
 
-LAMMPS is a popular particle-based simulation package used across many scientific fields. Two of its important file formats are the data file, which defines the simulation state, and the dump file, which records particle trajectories. These files are text-based and can be tedious to parse or manipulate. Python tools such as [Atomman](https://www.ctcms.nist.gov/potentials/atomman/) and [Pizza.py](https://github.com/lammps/pizza) support reading and writing LAMMPS particle-data files, but their interfaces can be complex. Other established simulation analysis software such as MDAnalysis [@michaud:jcc:2011; @gowers:pythonproc:2016], MDTraj [@mcgibbon:biophys:2015], and OVITO [@stukowski:msmse:2009] may only provide a subset of these features. As a result, researchers may rely on their own codes, but private codes are vulnerable to unnoticed errors and hinder reproducibility.
+LAMMPS is a popular particle-based simulation package used across many scientific fields. Two of its important file formats are the data file, which defines the simulation state, and the dump file, which records particle trajectories. These files are text-based and can be tedious to parse or manipulate. Python tools such as [Atomman](https://www.ctcms.nist.gov/potentials/atomman/) and [Pizza.py](https://github.com/lammps/pizza) support reading and writing LAMMPS particle-data files, but their interfaces can be complex. Moltemplate [@jewett:jmb:2021] can perform a variety of configuration tasks for LAMMPS simulations and includes Python tools for some file format conversions, but its Python interface is limited. Other established simulation analysis software such as MDAnalysis [@michaud:jcc:2011; @gowers:pythonproc:2016], MDTraj [@mcgibbon:biophys:2015], and OVITO [@stukowski:msmse:2009] also may only provide a subset of required features. As a result, researchers may rely on their own codes, but private codes are vulnerable to unnoticed errors and hinder reproducibility.
 
 lammpsio addresses this need by providing an object-oriented Python interface for reading and writing LAMMPS data and dump files while adhering to the TRUE (transparent, reproducible, usable by others, and extensible) principles of scientific software development [@thompson:molphys:2020]. It allows users to create a particle configuration (simulation input) and parse a trajectory (simulation output) in a consistent manner. lammpsio additionally supports interconversion with the GSD file used by HOOMD-blue, enabling sharing of particle configurations between simulation packages. This functionality is leveraged, for example, in our relative-entropy minimization software [@sreenivasan:jctc:2024] to support runtime selection of LAMMPS or HOOMD-blue as a simulation engine. Taken together, the features of lammpsio help meet needs for creating reproducible, interoperable, and maintainable simulation workflows for LAMMPS.
 
@@ -45,7 +45,7 @@ The DumpFile reads and writes LAMMPS dump files, which are text-based files that
 
 # Discussion
 
[…]
