+++ paper/paper.bib
@@ -16,6 +16,17 @@
 	pages        = {98–105},
 	doi          = {10.25080/Majora-629e541a-00e}
 }
+@article{jewett:jmb:2021,
+    title       = {Moltemplate: A Tool for Coarse-Grained Modeling of Complex Biological Matter and Soft Condensed Matter Physics},
+    author      = {Jewett, Andrew I. and Stelter, David and Lambert, Jason and Saladi, Shyam M. and Roscioni, Otello M. and Ricci, Matteo and Autin, Ludovic and Maritan, Martina and Bashusqeh, Saeed M. and Keyes, Tom and Dame, Remus T. and Shea, Joan-Emma and Jensen, Grant J. and Goodsell, David S.},
+    year        = 2021,
+    month       = {May},
+    journal     = {Journal of Molecular Biology},
+    volume      = 433,
+    number      = 11,
+    pages       = {166841},
+    doi         = {10.1016/j.jmb.2021.166841}
+}
 @article{mcgibbon:biophys:2015,
 	title        = {MDTraj: A Modern Open Library for the Analysis of Molecular Dynamics Trajectories},
 	author       = {McGibbon, Robert T. and Beauchamp, Kyle A. and Harrigan, Matthew P. and Klein, Christoph and Swails, Jason M. and Hernández, Carlos X. and Schwantes, Christian R. and Wang, Lee-Ping and Lane, Thomas J. and Pande, Vijay S.},
