### 2025-11-20 158079a
Revise tutorial cards
