+++ doc/source/tutorials.rst
@@ -15,25 +15,26 @@ Tutorials
         :link: ./tutorials/dimer_lattice_tutorial
         :link-type: doc
 
-        Use Snapshot to create and modify LAMMPS data files,
-        and convert them to HOOMD-blue's GSD format.
+        Create a LAMMPS data file and also convert it to HOOMD-blue's GSD
+        format.
 
     .. grid-item-card:: Analyzing a Lennard-Jones Fluid
         :link: ./tutorials/analysis_tutorial/analysis_tutorial
         :link-type: doc
 
-        Read and analyze a LAMMPS dump file to study the Lennard-Jones fluid
-        using the freud Python package.
+        Read a LAMMPS dump file for a Lennard-Jones fluid and analyze it using
+        ``freud``.
 
     .. grid-item-card:: Initializing an Ethanol-Water Mixture
         :link: ./tutorials/ethanol-water-tutorial/ethanol_water_mixture
         :link-type: doc
 
-        Use Snapshot to create a LAMMPS data file for an ethanol-water mixture.
+        Create a LAMMPS data file for an atomistic simulation of ethanol and
+        water using PACKMOL and LAMMPS molecule templates.
 
-    .. grid-item-card:: Analysis speedup using Numba
+    .. grid-item-card:: Speeding Up Analysis
         :link: ./tutorials/numba_tutorial/numba_tutorial
         :link-type: doc
 
-        Read and analyze a LAMMPS dump file to study a 100-mer polymer
-        using the NumPy and Numba library for accelerated computation.
+        Accelerate the analysis of a LAMMPS dump file for a linear polymer using
+        NumPy and Numba.
