### 2025-11-20 6c0d315
Numba tutorial (#95)
* Add tutorial and additional files required to run the simulation

* Pre-commit and documentation fix

* Change tutorial title

* Minor edits

* Edit numba tutorial.

* Remove initialization script and trajectory file

* Fix documentation error and minor changes to numba tutorial

* Minor fixes

* Update LAMMPS input and tutorial files.

* Remove unnecessary line break in tutorials.rst

* Include multiprocessing example for reading snapshots in parallel

* Fix documentation error

* Pre-commit fix

* Remove unnecessary imports and adjust multiprocessing worker count in du
[…]
