### 2025-12-08 mphowardlab/lammpsio#93 @mhlbapat
Looks good to me. I like the ethanol+water and numba tutorials. Can you please also update them in your documentation page? As far as the JOSS review is concerned, I would consider this issue to be closed.
