+++ src/lammpsio/__init__.py
@@ -4,4 +4,4 @@ from .dump import DumpFile
 from .snapshot import Snapshot
 from .topology import Angles, Bonds, Dihedrals, Impropers, LabelMap
 
-__version__ = "0.8.0"
+__version__ = "0.9.0"
