### 2025-12-10 #8962 @rkurchin
Putting on my AEiC hat to do some final checks:
- [ ] Please add "United States of America" to end of affiliation line
- [ ] Allen and Tildesley has a DOI, please add that: 10.1093/oso/9780198803195.001.0001
- [ ] The paper is on the long side for JOSS (we typically target 1000 words or less), and much of the content in "Design and implementation" seems like it could fit well in the docs (at least at a skim, I don't see all of this exposition there anywhere) and be present in a more compressed form here
- [ ] missing a period at the end of the "Availability" section
