### 2025-12-11 6c42471
Fix the missing period at end of Availability.
