### 2025-12-11 c148841
Tighten Design and implementation section.
