+++ paper/paper.md
@@ -35,13 +35,13 @@ lammpsio addresses this need by providing an object-oriented Python interface fo
 
 # Design and implementation
 
-lammpsio is organized around a central Snapshot object representing a single configuration of particles. Two other objects, the DataFile and DumpFile, provide read--write functionality for Snapshots in the corresponding LAMMPS file formats.
+lammpsio is organized around a central Snapshot object representing a single configuration of particles. DataFile and DumpFile objects provide read--write functionality for Snapshots in the corresponding LAMMPS file formats.
 
-The Snapshot stores information about a certain number of particles at a given time. The particles occupy a volume prescribed by a Box object, defined using the LAMMPS convention for restricted triclinic boxes. Each Snapshot stores per-particle data, such as position or velocity, in NumPy arrays. Topology data, defining the connectivity between subsets of particles (bonds, angles, dihedrals, and impropers) [@allen:2017], is also stored in the Snapshot using correspondingly named objects that manage per-connection data as NumPy arrays. Importantly, lammpsio allocates its per-particle and per-connection arrays with the correct shape and type only when they are first accessed. This design prevents mismatches in array attributes and can reduce memory requirements. The use of NumPy arrays additionally allows easy integration with the scientific Python ecosystem. The currently supported particle and topology information accommodates many common use cases; however, additional data can be attached to the Snapshot by subclassing or patching.
+The Snapshot stores information about particles at a given time in a volume prescribed using the LAMMPS restricted-triclinic convention. Each Snapshot stores per-particle data, such as position or velocity, in NumPy arrays. Topology data, defining the connectivity between subsets of particles [@allen:2017], is also stored in the Snapshot as NumPy arrays. Importantly, lammpsio allocates its per-particle and per-connection arrays with the correct shape and type only when they are first accessed. This design prevents mismatches in array attributes and can reduce memory requirements.
 
-The DataFile reads and writes LAMMPS data files, which are text-based files that store a single particle configuration. A LAMMPS data file is organized into a header section, which typically has high-level information about the simulation state, and a body containing multiple sections, many of which define per-particle and per-connection information for the configuration. Sections have standardized formats, but the Atoms section may have a different format depending on the atom style in the simulation. The DataFile's read method parses a data file into a Snapshot by tokenizing the headers and sections; the style of the Atoms section is inferred from a comment if present and may be manually specified otherwise. The DataFile's create method writes a data file containing only explicitly set information in the Snapshot and documented defaults for any data that is required but not explicitly set. Users can specify an atom style or allow lammpsio to infer the minimal style needed to represent their particles.
+The DataFile reads and writes LAMMPS data files, which are text-based files that store a single particle configuration. A data file contains a header with high-level simulation information and a body with sections defining per-particle and per-connection data. Sections have standardized formats, but the Atoms section may have a different format depending on the atom style in the simulation. The DataFile's read method parses a data file into a Snapshot by tokenizing the headers and sections. The DataFile's create method writes a data file containing only explicitly set information in the Snapshot and documented defaults for any data that is required but not explicitly set.
 
[…]
