### 2025-12-12 8485dae
Implement editorial changes.
