+++ paper/paper.md
@@ -35,13 +35,13 @@ lammpsio addresses this need by providing an object-oriented Python interface fo
 
 # Design and implementation
 
-lammpsio is organized around a central Snapshot object representing a single configuration of particles. DataFile and DumpFile objects provide read--write functionality for Snapshots in the corresponding LAMMPS file formats.
+lammpsio is organized around a central `Snapshot` object representing a single configuration of particles. `DataFile` and `DumpFile` objects provide read--write functionality for Snapshots in the corresponding LAMMPS file formats.
 
-The Snapshot stores information about particles at a given time in a volume prescribed using the LAMMPS restricted-triclinic convention. Each Snapshot stores per-particle data, such as position or velocity, in NumPy arrays. Topology data, defining the connectivity between subsets of particles [@allen:2017], is also stored in the Snapshot as NumPy arrays. Importantly, lammpsio allocates its per-particle and per-connection arrays with the correct shape and type only when they are first accessed. This design prevents mismatches in array attributes and can reduce memory requirements.
+The `Snapshot` stores information about particles at a given time in a volume prescribed using the LAMMPS restricted-triclinic convention. Each `Snapshot` stores per-particle data, such as position or velocity, in NumPy [@harris2020array] arrays. Topology data, defining the connectivity between subsets of particles [@allen:2017], is also stored in the `Snapshot` as NumPy arrays. Importantly, lammpsio allocates its per-particle and per-connection arrays with the correct shape and type only when they are first accessed. This design prevents mismatches in array attributes and can reduce memory requirements.
 
-The DataFile reads and writes LAMMPS data files, which are text-based files that store a single particle configuration. A data file contains a header with high-level simulation information and a body with sections defining per-particle and per-connection data. Sections have standardized formats, but the Atoms section may have a different format depending on the atom style in the simulation. The DataFile's read method parses a data file into a Snapshot by tokenizing the headers and sections. The DataFile's create method writes a data file containing only explicitly set information in the Snapshot and documented defaults for any data that is required but not explicitly set.
+The `DataFile` reads and writes LAMMPS data files, which are text-based files that store a single particle configuration. A data file contains a header with high-level simulation information and a body with sections defining per-particle and per-connection data. Sections have standardized formats, but the Atoms section may have a different format depending on the atom style in the simulation. The `DataFile`'s `read` method parses a data file into a `Snapshot` by tokenizing the headers and sections. The `DataFile`'s `create` method writes a data file containing only explicitly set information in the `Snapshot` and documented defaults for any data that is required but not explicitly set.
 
-The DumpFile reads and writes LAMMPS dump files, which are text-based files that store a trajectory (sequence of particle configurations). A dump file has a significantly more flexible format than a data file. It is organized into a sequence of items, including timestep, number of atoms (particles), box bounds, and atoms. The atoms item contains one or more columns representing per-particle properties and can be configured in a variety of ways. The DumpFile's read method loads one Snapshot at a time; users may provide a Snapshot to copy information not stored in the dump file (e.g., topology or omitted static properties). The DumpFile's create method generates a dump file from a sequence of Snapshots according to a user-defined schema for the atoms item. Dump files may have binary or compressed formats; gzip and Zstandard compression is currently supported.
[…]
