### 2025-12-13 8566c24
Add numpy citation.
