### 2025-12-22 a0dc2b4
Add citation information to doc and README.
