+++ doc/source/index.rst
@@ -28,3 +28,11 @@ in a variety of ways that improve workflow efficiency and reproducibility.
 
     ./credits
     ./license
+
+.. rubric:: Citing lammpsio
+
+When using ``lammpsio`` in published work, please cite:
+
+C. L. Petix, M. Kundu, and M. P. Howard., lammpsio: Transparent and
+reproducible handling of LAMMPS particle data in Python. *Journal of Open Source
+Software*, **10(116)**, 8962 (2025). `doi:10.21105/joss.08962 <https://doi.org/10.21105/joss.08962>`_
